+++ src/cell.c
@@ -40,8 +40,8 @@ int cll_allocate(CellLinkedGrid *cll, const Grid *grid) {
 }
 
 int cll_check_overlap(const Particle *p1, CellLinkedGrid *cll, const Grid* grid) {
-    unsigned int x_idx = (unsigned int) (p1->x / cll->s_x);
-    unsigned int y_idx = (unsigned int) (p1->y / cll->s_x);
+    long x_idx = (long) (p1->x / cll->s_x);
+    long y_idx = (long) (p1->y / cll->s_x);
     for (int i = -1; i < 2; i++) {
         for (int j = -1; j < 2; j++) {
             size_t idx = (x_idx + i + cll->n_x) % cll->n_x;
@@ -93,11 +93,11 @@ int random_gen(const char* filename, Grid *grid) {
 
 int square_gen(const char* filename, Grid *grid) {
     grid_allocate(grid);
-    unsigned int N_x = (unsigned int) (grid->Lx / grid->size) + 1;
-    unsigned int N_y = (unsigned int) (grid->Ly / grid->size) + 1;
+    long N_x = (long) (grid->Lx / grid->size) + 1;
+    long N_y = (long) (grid->Ly / grid->size) + 1;
     grid->N = N_x * N_y;
-    for (unsigned int x_idx = 0; x_idx < N_x; x_idx++) {
-        for (unsigned int y_idx = 0; y_idx < N_y; y_idx++) {
+    for (long x_idx = 0; x_idx < N_x; x_idx++) {
+        for (long y_idx = 0; y_idx < N_y; y_idx++) {
             grid->points[x_idx * N_y + y_idx].x = x_idx * grid->size;
             grid->points[x_idx * N_y + y_idx].y = y_idx * grid->size;
             grid->points[x_idx * N_y + y_idx].qw = 1;
@@ -110,12 +110,12 @@ int square_gen(const char* filename, Grid *grid) {
 
 // TODO: fix slight boundary overstep
 int hexagonal_gen(const char* filename, Grid *grid) {
-    unsigned int N_x = (unsigned int) ceil(grid->Lx / grid->size / (sqrt(3)/2));
-    unsigned int N_y = (unsigned int) (grid->Ly / grid->size) + 1;
+    long N_x = (long) ceil(grid->Lx / grid->size / (sqrt(3)/2));
+    long N_y = (long) (grid->Ly / grid->size) + 1;
     grid-> N = N_x * N_y;
     grid_allocate(grid);
-    for (unsigned int x_idx = 0; x_idx < N_x; x_idx++) {
-        for (unsigned int y_idx = 0; y_idx < N_y; y_idx++) {
+    for (long x_idx = 0; x_idx < N_x; x_idx++) {
+        for (long y_idx = 0; y_idx < N_y; y_idx++) {
             grid->points[x_idx * N_y + y_idx].x = x_idx * grid->size * sqrt(3)/2;
             grid->points[x_idx * N_y + y_idx].y = y_idx * grid->size;
             grid->points[x_idx * N_y + y_idx].qw = 1;
